Return `None` on incompatible wire types
Parsed fields with unexpected wire types should be handled the same as
unknown fields. This is the behavior of the Google C++ protobuf
implementation. Avoid returning `Error` on incompatible wire types so
that this case may handled differently from genuine errors. For getters
of non-repeated fields, use `Option` for the return type and return
`None` when the wire type is incompatible. For getters of repeated
fields, simply return an iterator that yields nothing instead of
returning an error.